fix tet suite for reference index
